+++ CHANGELOG.md
@@ -10,7 +10,7 @@ and this project adheres to
 
 ## Added
 
-- Add `join_and` template tags
+- Add `join_and` and `list_key` template tags
 - Add address and new properties on client OrganizationSerializer
 - Add admin endpoints to create/update/delete organization addresses
 - Add several admin api endpoint filters